+++ extras/tests/JsonArray/isNull.cpp
@@ -6,29 +6,53 @@
 #include <catch.hpp>
 
 TEST_CASE("JsonArray::isNull()") {
-  DynamicJsonDocument doc(4096);
-
   SECTION("returns true") {
     JsonArray arr;
     REQUIRE(arr.isNull() == true);
   }
 
   SECTION("returns false") {
+    DynamicJsonDocument doc(4096);
     JsonArray arr = doc.to<JsonArray>();
     REQUIRE(arr.isNull() == false);
   }
 }
 
 TEST_CASE("JsonArrayConst::isNull()") {
-  DynamicJsonDocument doc(4096);
-
   SECTION("returns true") {
     JsonArrayConst arr;
     REQUIRE(arr.isNull() == true);
   }
 
   SECTION("returns false") {
+    DynamicJsonDocument doc(4096);
     JsonArrayConst arr = doc.to<JsonArray>();
     REQUIRE(arr.isNull() == false);
+  }
+}
+
+TEST_CASE("JsonArray::operator bool()") {
+  SECTION("returns false") {
+    JsonArray arr;
+    REQUIRE(static_cast<bool>(arr) == false);
+  }
+
+  SECTION("returns true") {
+    DynamicJsonDocument doc(4096);
+    JsonArray arr = doc.to<JsonArray>();
+    REQUIRE(static_cast<bool>(arr) == true);
+  }
+}
+
+TEST_CASE("JsonArrayConst::operator bool()") {
+  SECTION("returns false") {
+    JsonArrayConst arr;
+    REQUIRE(static_cast<bool>(arr) == false);
+  }
+
+  SECTION("returns true") {
+    DynamicJsonDocument doc(4096);
+    JsonArrayConst arr = doc.to<JsonArray>();
+    REQUIRE(static_cast<bool>(arr) == true);
   }
 }