+++ src/ArduinoJson/Array/ArrayRef.hpp
@@ -36,6 +36,10 @@ class ArrayRefBase {
     return _data == 0;
   }
 
+  FORCE_INLINE operator bool() const {
+    return _data != 0;
+  }
+
   FORCE_INLINE size_t memoryUsage() const {
     return _data ? _data->memoryUsage() : 0;
   }