+++ CHANGELOG.md
@@ -5,6 +5,7 @@ HEAD
 ----
 
 * Added support for custom writer classes (issue #1088)
+* Added conversion from `JsonArray` and `JsonObject` to `bool`, to be consistent with `JsonVariant`
 
 v6.12.0 (2019-09-05)
 -------